chore: Set up repository (.gitignore, .gitmodules, .gitattributes, CMake project structure)
Added googletest submodule and created test directory with a first file to add different
uni tests. In addition, a CMakeLists.txt has been created to init the project with the name
of "ekf_mono_slam" and included test directory (more utest .cpp files are left to be added).
Finally, a Makefile to easily build the project.

Makefile available commands:

- make clean
- make build
- make debug
- make format (still left to add clang formatter)
- make all